+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.1 / 2019-07-24
+
+* Enhancements
+
+  * Support Amazon STS with Secretfile.group {}
+  * Support dashes in vault paths
+
 1.0.0 / 2019-07-24
 
 Initial release - inspired by https://github.com/erithmetic/secret_garden